requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.40.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Vision models
llava-torch>=1.0.0
timm>=0.9.0
einops>=0.7.0

# Data processing
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.0
omegaconf>=2.3.0
huggingface-hub>=0.20.0

# Optional: Video processing
decord>=0.6.0
av>=10.0.0

# Optional: Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Optional: Graph processing
torch-geometric>=2.4.0
networkx>=3.0
